feat: implement core budget contract logic with multi-asset support, category management, and spending protection
