+++ contracts/currency-conversion/src/test.rs
@@ -1,0 +1,3 @@
+#![cfg(test)]
+
+use super::*;
